Booking Time Off

+++ src/main/java/com/mgmtp/repository/EmployeeRepository.java
@@ -3,9 +3,6 @@ package com.mgmtp.repository;
 import com.mgmtp.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-/**
- * Created by Tan Dat on 18/12/2016.
- */
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
     Employee findById(int employeeId);
 }
